requirements.txt
Flask==2.3.3
Pillow==10.0.1
pandas==2.1.1
google-generativeai==0.3.2
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0